Some simple perf test implemetation + separate file.
